+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
-- Add `causation_id` to events
+- [**BC**] Make event fields type safe
+- [**BC**] Add `causation_id` to events
 
 ## 3.2.0 - 2019-02-27
 - Add `consumeStreamToOffset` function to `Consumer`